python-ml-control/requirements.txt
wheel
picamera2
rpi-libcamera
numpy>=1.26.0, <2.0.0
pyzmq